Further update API UX
There are two semi-major changes, both with the purpose of improving the
user interface of the general C3D API.

The Parameter type(s) were refined after a careful re-reading of the C3D
online documentation revealed three major 'types' of parameters:

- Scalar
  + Examples include POINT:RATE, ANALOG:USED, etc.
- Arrays of Strings
  + Examples include any LABEL or DESCRIPTION parameters
- Arrays of `Type{T}`
  + One example being FORCE_PLATFORM:CAL_MATRIX

New Parameter types were implemented for better usability of these three
main parameter types.

In addition, a `getproperty(g::Group, name)` method was added that
allows shorter (and more importantly distinct) way of getting the data
contained in a Parameter. These two changes in combination have the
benefit of not being required to index in a single element array, which
although is not a huge change in terms of the number of characters
typed, this is much more readable (less noise).